Move ProxyStatus to serve/schema.py
